pyproject.toml
[project]
name = "pyridine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "frozendict>=2.4.7",
    "kiteconnect>=5.0.1",
    "pytest>=9.0.2",
    "python-dotenv-vault>=0.7.0",
    "requests>=2.32.5",
    "matplotlib>=3.8",
]
